+++ Tarea 4 - Ventaneo/Procesos/archivos.h
@@ -1,5 +1,5 @@
 #ifndef ARCHIVOS_H
 #define ARCHIVOS_H
-void leerArchivo(int* datos, char* nombreArchivo);
-void escribirArchivo(int* datos);
+void leerArchivo(double* datos, char* nombreArchivo);
+void escribirArchivo(double* datos);
 #endif
