+++ Tarea 4 - Ventaneo/Procesos/procesamiento.h
@@ -1,4 +1,4 @@
 #ifndef PROCESAMIENTO_H
 #define PROCESAMIENTO_H
-void ventanaHann(int* datos);
+void ventanaHann(double* datos);
 #endif
